taf-ms/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
psycopg2-binary
apscheduler
python-multipart
PyYAML
